requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-Bcrypt
Flask-JWT-Extended
python-dotenv
mysql
pymysql
flask-login
